+++ docs/CHANGES.md
@@ -23,7 +23,7 @@ Full implementation of branched network topologies.
 Overhaul of network creation.
 
 - Sequential layer addition. Branches are user-added.
-    - Rationale: PyTorch-style custom network definition didn't work. Operators go out of scope when the `forward` method is called, so the network cannot access the operators in the backward pass
+    - Rationale: PyTorch-style custom network definition didn't work. Operators go out of scope when the `forward` method is called, so the network cannot access the operators in the backward pass.
     - Each layer is stored in a std::vector. Layers track their predecessor and successor index upon addition.
     - Branches are special layers with multiple predecessors or successors.
 - Tensor objects (as used in v0.5) are no longer used. The user's data storage is the xt::xarray.